+++ examples/list-servers.rs
@@ -24,7 +24,7 @@ fn main() {
     let identity = openstack::auth::Identity::from_env()
         .expect("Failed to create an identity provider from the environment");
     let session = openstack::Session::new(identity);
-    let servers = openstack::compute::servers::manager(&session).list()
+    let servers = openstack::compute::v2(&session).servers().list()
         .expect("Cannot list servers");
     for s in &servers {
         println!("ID = {}, Name = {}", s.id(), s.name());